feat(PolicyDetails): code refactoring
