+++ src/SmartComponents/EditPolicy/hooks/index.js
@@ -1,2 +1,3 @@
 import { useState } from 'react';
+
 export const useNewRulesAlertState = (...args) => useState(...args);
